+++ cmd/hidden_lake/helpers/encryptor/pkg/app/init_app.go
@@ -14,7 +14,7 @@ import (
 )
 
 // initApp work with the raw data = read files, read args
-func InitApp(pArgs []string, pDefaultPath, pDefaultKey string, pDefaultParallel uint64) (types.IRunner, error) {
+func InitApp(pArgs []string, pDefaultPath, pDefaultPrivPath string, pDefaultParallel uint64) (types.IRunner, error) {
 	strParallel := flag.GetFlagValue(pArgs, "parallel", strconv.FormatUint(pDefaultParallel, 10))
 	setParallel, err := strconv.Atoi(strParallel)
 	if err != nil {
@@ -30,8 +30,8 @@ func InitApp(pArgs []string, pDefaultPath, pDefaultKey string, pDefaultParallel 
 		return nil, utils.MergeErrors(ErrInitConfig, err)
 	}
 
-	inputKey := flag.GetFlagValue(pArgs, "key", pDefaultKey)
-	privKey, err := initapp.GetPrivKey(inputKey, cfg.GetSettings().GetKeySizeBits())
+	privPath := flag.GetFlagValue(pArgs, "priv", pDefaultPrivPath)
+	privKey, err := initapp.GetPrivKey(privPath, cfg.GetSettings().GetKeySizeBits())
 	if err != nil {
 		return nil, utils.MergeErrors(ErrGetPrivateKey, err)
 	}